Update
Functionality overhaul
Wanting to satisfy multiple different quality of life features was causing me to create messy code.
This was cause because i didnt have a clear vision of what i wanted.
Now i have a clear vision of what i want so now it should be easier to finish this project